Rsod wheels

+++ wetb/signal/filters/cy_filters.py
@@ -14,7 +14,7 @@ import numpy as np
 @cython.locals(alpha=cython.float, i=cython.int)
 def cy_low_pass_filter(inp, delta_t, tau):  #cpdef cy_low_pass_filter(np.ndarray[double,ndim=1] inp, double delta_t, double tau):
     #cdef np.ndarray[double,ndim=1] output
-    output = np.empty_like(inp, dtype=np.float)
+    output = np.empty_like(inp, dtype=np.float32)
     output[0] = inp[0]
 
     alpha = delta_t / (tau + delta_t)
@@ -27,7 +27,7 @@ def cy_dynamic_low_pass_filter(inp, delta_t, tau, method=1):  #cpdef cy_dynamic_
     #cdef np.ndarray[double,ndim=1] output, alpha
     #cdef int i
 
-    output = np.empty_like(inp, dtype=np.float)
+    output = np.empty_like(inp, dtype=np.float32)
     output[0] = inp[0]
 
     if method == 1:
@@ -46,7 +46,7 @@ def cy_dynamic_low_pass_filter_2d(inp, delta_t, tau, method=1):  #cpdef cy_dynam
     #cdef np.ndarray[double,ndim=2] output, alpha
     #cdef int i
 
-    output = np.empty_like(inp, dtype=np.float)
+    output = np.empty_like(inp, dtype=np.float32)
     output[0] = inp[0]
 
     if method == 1:
@@ -64,7 +64,7 @@ def cy_dynamic_low_pass_filter_2d(inp, delta_t, tau, method=1):  #cpdef cy_dynam
 def cy_dynamic_low_pass_filter_test(inp):  #cpdef cy_dynamic_low_pass_filter_test(np.ndarray[double,ndim=2] inp):
     #cdef np.ndarray[double,ndim=2] output, alpha
     #cdef int i
-    output = np.empty_like(inp, dtype=np.float)
+    output = np.empty_like(inp, dtype=np.float32)
     output[0] = inp[0]
     for i in range(1, inp.shape[0]):
         output[i] = inp[i]
@@ -74,7 +74,7 @@ def cy_dynamic_low_pass_filter_test(inp):  #cpdef cy_dynamic_low_pass_filter_tes
 @cython.locals(alpha=cython.float, i=cython.int)
 def cy_high_pass_filter(inp, delta_t, tau):  #cpdef cy_high_pass_filter(np.ndarray[double,ndim=1] inp, double delta_t, double tau):
     #cdef np.ndarray[double,ndim=1] output
-    output = np.empty_like(inp, dtype=np.float)
+    output = np.empty_like(inp, dtype=np.float32)
     output[0] = inp[0]
     alpha = tau / (tau + delta_t)
     for i in range(1, inp.shape[0]):
